+++ examples/uom.rs
@@ -12,8 +12,8 @@ fn main() {
     let f_x = array![Power::new::<kilowatt>(0.25), Power::new::<kilowatt>(0.75)];
     // `uom::si::Quantity` is repr(transparent), meaning it has the same memory layout as its contained type.
     // This means we can get the contained type via transmuting.
-    let interp: Interp1DViewed<&f64, _> = unsafe {
-        Interp1D::new(
+    let interp: Interp1DView<&f64, _> = unsafe {
+        Interp1DView::new(
             std::mem::transmute::<ArrayView1<Ratio>, ArrayView1<f64>>(x.view()),
             std::mem::transmute::<ArrayView1<Power>, ArrayView1<f64>>(f_x.view()),
             strategy::Linear,