+++ src/interpolator/three/strategies.rs
@@ -8,7 +8,7 @@ where
 {
     fn interpolate(
         &self,
-        data: &InterpData3D<D>,
+        data: &InterpData3DBase<D>,
         point: &[D::Elem; 3],
     ) -> Result<D::Elem, InterpolateError> {
         // Extrapolation is checked previously in Interpolator::interpolate,
@@ -164,7 +164,7 @@ where
     D::Elem: Float + Debug,
 {
     /// Ensures all grid dimensions are uniformly spaced.
-    fn validate(&self, data: &InterpData3D<D>) -> Result<(), ValidateError> {
+    fn validate(&self, data: &InterpData3DBase<D>) -> Result<(), ValidateError> {
         check_uniform_grid(data.grid[0].view(), 0)?;
         check_uniform_grid(data.grid[1].view(), 1)?;
         check_uniform_grid(data.grid[2].view(), 2)
@@ -172,7 +172,7 @@ where
 
     fn interpolate(
         &self,
-        data: &InterpData3D<D>,
+        data: &InterpData3DBase<D>,
         point: &[D::Elem; 3],
     ) -> Result<D::Elem, InterpolateError> {
         let x_step = data.grid[0][1] - data.grid[0][0];
@@ -213,7 +213,7 @@ where
 {
     fn interpolate(
         &self,
-        data: &InterpData3D<D>,
+        data: &InterpData3DBase<D>,
         point: &[D::Elem; 3],
     ) -> Result<D::Elem, InterpolateError> {
         // x
@@ -256,7 +256,7 @@ where
     D::Elem: PartialOrd + Copy + Debug,
 {
     /// Ensures the number of provided step directions matches the interpolator dimensionality.
-    fn validate(&self, _data: &InterpData3D<D>) -> Result<(), ValidateError> {
+    fn validate(&self, _data: &InterpData3DBase<D>) -> Result<(), ValidateError> {
         if self.0.len() != 1 && self.0.len() != 3 {
             return Err(ValidateError::Other(format!(
                 "Step strategy has {} directions but interpolator is 3-D (expected 1 or 3)",
@@ -268,7 +268,7 @@ where
 
     fn interpolate(
         &self,
-        data: &InterpData3D<D>,
+        data: &InterpData3DBase<D>,
         point: &[D::Elem; 3],
     ) -> Result<D::Elem, InterpolateError> {
         let i = locate_step_index(self.dir(0), data.grid[0].view(), &point[0]);
@@ -290,7 +290,7 @@ where
 {
     fn interpolate(
         &self,
-        data: &InterpData3D<D>,
+        data: &InterpData3DBase<D>,
         point: &[D::Elem; 3],
     ) -> Result<D::Elem, InterpolateError> {
         let i = locate_step_index(StepDirection::Lower, data.grid[0].view(), &point[0]);
@@ -311,7 +311,7 @@ where
 {
     fn interpolate(
         &self,
-        data: &InterpData3D<D>,
+        data: &InterpData3DBase<D>,
         point: &[D::Elem; 3],
     ) -> Result<D::Elem, InterpolateError> {
         let i = locate_step_index(StepDirection::Upper, data.grid[0].view(), &point[0]);